Move TrackerServer, SeedersTracker to torrent.tracker package
