+++ src/main/java/net/yeputons/spbau/spring2016/torrent/client/ConsoleClient.java
@@ -3,7 +3,7 @@ package net.yeputons.spbau.spring2016.torrent.client;
 import net.yeputons.spbau.spring2016.torrent.FileDescription;
 import net.yeputons.spbau.spring2016.torrent.StateFileHolder;
 import net.yeputons.spbau.spring2016.torrent.TorrentConnection;
-import net.yeputons.spbau.spring2016.torrent.TrackerServer;
+import net.yeputons.spbau.spring2016.torrent.tracker.TrackerServer;
 import net.yeputons.spbau.spring2016.torrent.protocol.FileEntry;
 import net.yeputons.spbau.spring2016.torrent.protocol.ListRequest;
 import net.yeputons.spbau.spring2016.torrent.protocol.UploadRequest;
