+++ src/main/java/net/yeputons/spbau/spring2016/torrent/tracker/TrackerServer.java
@@ -1,5 +1,9 @@
-package net.yeputons.spbau.spring2016.torrent;
+package net.yeputons.spbau.spring2016.torrent.tracker;
 
+import net.yeputons.spbau.spring2016.torrent.SocketDataStreamsWrapper;
+import net.yeputons.spbau.spring2016.torrent.StateFileHolder;
+import net.yeputons.spbau.spring2016.torrent.StateHolder;
+import net.yeputons.spbau.spring2016.torrent.StateMemoryHolder;
 import net.yeputons.spbau.spring2016.torrent.protocol.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
